Refactor vNode
