+++ src/models/device.js
@@ -60,10 +60,10 @@ export const VALIDATOR_STATUS = {
 export const MAX_ERROR_COUNT = 5;
 
 export default class Device {
-  static CODE_STOP = DEVICE_STATUS.CODE_STOP;
+  static CODE_STOP    = DEVICE_STATUS.CODE_STOP;
   static CODE_PENDING = DEVICE_STATUS.CODE_PENDING;
-  static CODE_START = DEVICE_STATUS.CODE_START;
-  static CODE_MINING = DEVICE_STATUS.CODE_MINING;
+  static CODE_START   = DEVICE_STATUS.CODE_START;
+  static CODE_MINING  = DEVICE_STATUS.CODE_MINING;
   static CODE_SYNCING = DEVICE_STATUS.CODE_SYNCING;
 
   constructor(data){
