+++ src/screens/Node/Node.constant.js
@@ -1,33 +1,24 @@
-
-export const ACTION_FETCHING_NODES_INFO
-  = '[NODE]ACTION_FETCHING_NODES_INFO';
-
-export const ACTION_FETCH_NODES_INFO_FAIL
-  = '[NODE]ACTION_FETCH_NODES_INFO_FAIL';
-
-export const ACTION_FETCHED_NODES_INFO
-  = '[NODE]ACTION_FETCHED_NODES_INFO';
-
+//new flow
 export const ACTION_UPDATE_LIST_NODE_DEVICE
   = '[NODE]ACTION_UPDATE_LIST_NODE_DEVICE';
 
-export const ACTION_UPDATE_FETCHING
-  = '[NODE]ACTION_UPDATE_FETCHING';
+export const ACTION_FETCHING_NODES_INFO_FROM_API
+  = '[NODE]ACTION_FETCHING_NODES_INFO_FROM_API';
 
-export const ACTION_FETCHING_FULL_INFO
-  = '[NODE]ACTION_FETCHING_FULL_INFO';
+export const ACTION_FETCH_NODES_INFO_FROM_API_FAIL
+  = '[NODE]ACTION_FETCH_NODES_INFO_FROM_API_FAIL';
 
-export const ACTION_FETCH_FULL_INFO_FAIL
-  = '[NODE]ACTION_FETCH_FULL_INFO_FAIL';
+export const ACTION_FETCHED_NODES_INFO_API
+  = '[NODE]ACTION_FETCHED_NODES_INFO_API';
 
-export const ACTION_FETCHED_FULL_INFO
-  = '[NODE]ACTION_FETCHED_FULL_INFO';
+export const ACTION_SET_TOTAL_VNODE
+  = '[NODE]ACTION_SET_TOTAL_VNODE';
 
-export const ACTION_UPDATE_WITH_DRAWING
-  = '[NODE]ACTION_UPDATE_WITH_DRAWING';
-
-export const ACTION_UPDATE_COMBINE_REWARDS
-  = '[NODE]ACTION_UPDATE_COMBINE_REWARDS';
+export const ACTION_UPDATE_NUMBER_LOADED_VNODE_BLS
+  = '[NODE]ACTION_UPDATE_NUMBER_LOADED_VNODE_BLS';
 
 export const ACTION_UPDATE_MISSING_SETUP
   = '[NODE]ACTION_UPDATE_MISSING_SETUP';
+
+export const ACTION_UPDATE_FETCHING
+  = '[NODE]ACTION_UPDATE_FETCHING';
