+++ src/screens/Node/Node.enhanceWelcome.js
@@ -1,0 +1,16 @@
+import React from 'react';
+import ErrorBoundary from '@src/components/ErrorBoundary';
+
+const enhance = WrappedComp => props => {
+  return (
+    <ErrorBoundary>
+      <WrappedComp
+        {...{
+          ...props,
+        }}
+      />
+    </ErrorBoundary>
+  );
+};
+
+export default enhance;
