+++ src/screens/Node/Node.reducer.js
@@ -2,15 +2,13 @@ import {persistReducer} from 'redux-persist';
 import AsyncStorage from '@react-native-community/async-storage';
 import autoMergeLevel2 from 'redux-persist/es/stateReconciler/autoMergeLevel2';
 import {
-  ACTION_FETCH_FULL_INFO_FAIL, ACTION_FETCH_NODES_INFO_FAIL,
-  ACTION_FETCHED_FULL_INFO, ACTION_FETCHED_NODES_INFO,
-  ACTION_FETCHING_FULL_INFO,
-  ACTION_FETCHING_NODES_INFO,
-  ACTION_UPDATE_COMBINE_REWARDS,
-  ACTION_UPDATE_FETCHING,
+  ACTION_FETCH_NODES_INFO_FROM_API_FAIL,
+  ACTION_FETCHED_NODES_INFO_API,
+  ACTION_FETCHING_NODES_INFO_FROM_API,
   ACTION_UPDATE_LIST_NODE_DEVICE,
   ACTION_UPDATE_MISSING_SETUP,
-  ACTION_UPDATE_WITH_DRAWING
+  ACTION_SET_TOTAL_VNODE,
+  ACTION_UPDATE_NUMBER_LOADED_VNODE_BLS, ACTION_UPDATE_FETCHING
 } from '@screens/Node/Node.constant';
 import {PRV} from '@services/wallet/tokenService';
 
@@ -27,7 +25,7 @@ const initAllTokens = [PRV];
 const initCombineRewards = {
   rewards:      null,
   withdrawable: false,
-  noRewards:    false
+  noRewards:    true
 };
 
 const initMissingSetup = {
@@ -35,21 +33,35 @@ const initMissingSetup = {
   verifyProductCode: ''
 };
 
+const initVNodeOptions = {
+  hasVNode:         true,
+  vNodeNotHaveBLS:  -1,
+};
+
 const initialState = {
-  isFetching: false,
-  withdrawing: false,
-  listDevice: [], //List node device
-  committees: initCommittees,
-  nodeRewards: {},
-  allTokens: initAllTokens,
+  // withdrawing:    false,
+  committees:     initCommittees,
+  allTokens:      initAllTokens,
   combineRewards: initCombineRewards,
-  missingSetup: initMissingSetup
+
+  //New Flow
+  isFetching:     false,
+  isFetched:      false,
+  isRefreshing:   false,
+  noRewards:      true,
+
+  listDevice:     [], //List nodACTION_FETCHING_NODES_INFO_FROM_APIe device
+  nodesFromApi:   [], //api support cached node info from Chain
+  vNodeOptions:   initVNodeOptions,
+  nodeRewards:    null,
+  missingSetup:   initMissingSetup,
 };
 
 const nodeReducer = (state = initialState, action) => {
   switch (action.type) {
+  // new flow
   case ACTION_UPDATE_LIST_NODE_DEVICE: {
-    const { listDevice, isFetching } = action?.payload;
+    let { listDevice, isFetching } = action?.payload;
     if (isFetching !== null && isFetching !== undefined) {
       state = {
         ...state,
@@ -58,55 +70,7 @@ const nodeReducer = (state = initialState, action) => {
     }
     return {
       ...state,
-      listDevice: listDevice || [],
-      withdrawing: false
-    };
-  }
-  case ACTION_UPDATE_FETCHING: {
-    const { isFetching } = action;
-    return {
-      ...state,
-      isFetching,
-    };
-  }
-  case ACTION_FETCHING_FULL_INFO: {
-    return {
-      ...state,
-      isFetching: true,
-    };
-  }
-  case ACTION_FETCH_FULL_INFO_FAIL: {
-    return {
-      ...state,
-      isFetching: false,
-    };
-  }
-  case ACTION_FETCHED_FULL_INFO: {
-    const { committees, nodeRewards, allTokens } = action?.payload;
-    return {
-      ...state,
-      isFetching: false,
-      committees,
-      nodeRewards,
-      allTokens
-    };
-  }
-  case ACTION_UPDATE_WITH_DRAWING: {
-    const { withdrawing } = action;
-    return {
-      ...state,
-      withdrawing
-    };
-  }
-  case ACTION_UPDATE_COMBINE_REWARDS: {
-    const { rewards, withdrawable, noRewards } = action?.payload;
-    return {
-      ...state,
-      combineRewards: {
-        rewards,
-        withdrawable,
-        noRewards
-      }
+      listDevice,
     };
   }
   case ACTION_UPDATE_MISSING_SETUP: {
@@ -119,24 +83,67 @@ const nodeReducer = (state = initialState, action) => {
       }
     };
   }
-  case ACTION_FETCHING_NODES_INFO: {
+  case ACTION_FETCHING_NODES_INFO_FROM_API: {
+    const { isRefresh } = action;
     return {
       ...state,
-      isFetching: true
+      isFetching: !isRefresh,
+      isRefreshing: isRefresh || false,
+      isFetched:  false
     };
   }
-  case ACTION_FETCH_NODES_INFO_FAIL: {
+  case ACTION_FETCH_NODES_INFO_FROM_API_FAIL: {
     return {
       ...state,
-      isFetching: false
+      isFetching:   false,
+      isFetched:    false,
+      isRefreshing: false
     };
   }
-  case ACTION_FETCHED_NODES_INFO: {
-    const { listNodes } = action?.payload;
+  case ACTION_FETCHED_NODES_INFO_API: {
+    const { nodesFromApi, listDevice, nodeRewards, noRewards } = action?.payload;
+
     return {
       ...state,
       isFetching: false,
-      listNodes
+      isRefreshing: false,
+      listDevice: listDevice || [],
+      nodesFromApi,
+      nodeRewards,
+      isFetched: true,
+      noRewards
+    };
+  }
+  case ACTION_SET_TOTAL_VNODE: {
+    const { hasVNode, vNodeNotHaveBLS } = action?.payload;
+    const { vNodeOptions }  = state;
+    return {
+      ...state,
+      vNodeOptions: {
+        ...vNodeOptions,
+        hasVNode,
+        vNodeNotHaveBLS
+      }
+    };
+  }
+  case ACTION_UPDATE_NUMBER_LOADED_VNODE_BLS: {
+    const { vNodeOptions }    = state;
+    const { vNodeNotHaveBLS } = vNodeOptions;
+    const result = vNodeNotHaveBLS - 1;
+    const _vNodeNotHaveBLS = result >= 0 ? result : 0;
+    return {
+      ...state,
+      vNodeOptions: {
+        ...vNodeOptions,
+        vNodeNotHaveBLS: _vNodeNotHaveBLS,
+      }
+    };
+  }
+  case ACTION_UPDATE_FETCHING: {
+    const { isFetching } = action;
+    return {
+      ...state,
+      isFetching,
     };
   }
   default:
