+++ src/screens/Node/Node.selector.js
@@ -3,25 +3,41 @@ import { createSelector } from 'reselect';
 export const nodeSelector = createSelector(
   (state) => state.node,
   (node) => {
-    const { rewards,withdrawable, noRewards  } = node?.combineRewards;
 
     return {
       ...node, 
       isFetching:   node?.isFetching,
       withdrawing:  node?.withdrawing,
-      listDevice:   node?.listDevice,
       allTokens:    node?.allTokens,
-      nodeRewards:  node?.nodeRewards,
-      committees:   node?.committees,
+      // nodeRewards:  node?.nodeRewards,
+      // committees:   node?.committees,
 
       missingSetup: node?.missingSetup,
 
       //Combine awards below
-      rewards:      rewards,
-      withdrawable: withdrawable,
-      noRewards:    noRewards,
+      rewards:      node?.rewards,
+      withdrawable: node?.withdrawable,
+
+      //new flow
+      listDevice:   node?.listDevice, //List node device
+      nodesFromApi: node?.nodesFromApi,
+      noRewards:    node?.noRewards,
     };
   }
+);
+
+export const vNodeOptionsSelector = createSelector(
+  nodeSelector,
+  (node) => {
+    const {
+      hasVNode,
+      vNodeNotHaveBLS,
+    } = node?.vNodeOptions;
+    return {
+      hasVNode,
+      vNodeNotHaveBLS,
+    };
+  },
 );
 
 export const missingSetupNodeSelector = createSelector(
