+++ src/screens/Node/Node.services.js
@@ -3,14 +3,18 @@ import { formatBodyGetNodesInfo } from '@screens/Node/Node.builder';
 
 export const apiGetNodesInfo = async () => {
   return new Promise(async (resolve, reject) => {
-    const body = await formatBodyGetNodesInfo();
-    return http
-      .post('pnode/get-node-info', body)
-      .then((res) => {
-        resolve(res);
-      })
-      .catch((error) => {
-        reject(error);
-      });
+    try {
+      const body = await formatBodyGetNodesInfo();
+      http
+        .post('pnode/get-node-info', body)
+        .then((res) => {
+          resolve(res);
+        })
+        .catch((error) => {
+          throw error;
+        });
+    } catch (error) {
+      reject(error);
+    }
   });
 };
