+++ src/screens/Node/components/NodeItem/index.js
@@ -1,0 +1,3 @@
+// eslint-disable-next-line import/no-cycle
+export { default } from './NodeItem';
+export { default as withNodeItemEnhance } from './NodeItem.enhance';
