+++ src/screens/Node/components/_NodeItem.js
