+++ src/screens/Node/components/style.js
@@ -184,4 +184,18 @@ export const rewardStyle = StyleSheet.create({
   },
 });
 
+export const nodeItemStyle = StyleSheet.create({
+  wrapperRaw: {
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+    alignContent: 'center'
+  },
+  btnStyle: {
+    width: 30,
+    alignItems: 'center',
+    alignContent: 'center',
+    justifyContent: 'center',
+  },
+});
+
 export default style;
